+++ RuboCop.novaextension/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 
+- Added an informative notice to the user when RuboCop could not be found.
 - Restructured the project to allow for npm dependencies to be consumed.
 - Added a new icon for the extension.
 - Added documentation to the README.